+++ assets/javascripts/kitten/components/form/form-actions.js
@@ -1,0 +1,12 @@
+import React from 'react'
+import classNames from 'classnames'
+
+export class FormActions extends React.Component {
+  render() {
+    return (
+      <div className={ classNames('k-FormActions', this.props.className) }>
+        { this.props.children }
+      </div>
+    )
+  }
+}